requirements.txt
Django==2.2.1
django-ace==1.0.5
django-crispy-forms==1.7.2
pytz==2019.1
sqlparse==0.3.0
